# Avoid union of identical geometries in summarize()